% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPAsample.R
\docType{data}
\name{DPAsample}
\alias{DPAsample}
\title{Simulated DPA Sample Data}
\format{A data frame with 11060 observations on the following 3 variables.
\describe{ \item{subject}{participant number}
\item{duration}{the duration of the fixation, in ms}
\item{condition}{1 = fast condition; 2 = slow condition} }}
\source{
Reingold, E. M. & Sheridan, H. (2014). Estimating the divergence
point: A novel distributional analysis procedure for determining the onset
of the influence of experimental variables. Frontiers in Psychology. doi:
10.3389/fpsyg.2014.01432.
}
\usage{
DPAsample
}
\description{
Simulated sample fixation duration data of one-hundred and four participants
as provided in Reingold and Sheridan (2014).
}
\keyword{datasets}

