\name{DPAsample}
\alias{DPAsample}
\docType{data}
\title{
Simulated DPA Sample Data
}
\description{
Simulated sample fixation duration data of one-hundred and four participants as provided in Reingold and Sheridan (2014).
}
\usage{data("DPAsample")}
\format{
  A data frame with 11060 observations on the following 3 variables.
  \describe{
    \item{\code{subject}}{participant number}
    \item{\code{duration}}{the duration of the fixation, in ms}
    \item{\code{condition}}{1 = fast condition; 2 = slow condition}
  }
}
\source{
Reingold, E. M. & Sheridan, H. (2014). Estimating the divergence point: A novel distributional analysis procedure for determining the onset of the influence of experimental variables. Frontiers in Psychology. doi: 10.3389/fpsyg.2014.01432.
}
\keyword{datasets}
