% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv.curv.R
\name{surv.curv}
\alias{surv.curv}
\title{survival curves}
\usage{
surv.curv(x, .bin, sample = FALSE, sample.n = length(x), replace = TRUE)
}
\arguments{
\item{x}{A numeric vector of RT data.}

\item{.bin}{A numeric vector specifying upper bounds of each timing bin}

\item{sample}{logical. If \code{TRUE}, random sample of \code{x} is used to
calculate survival proportion.}

\item{sample.n}{A size of the sample when \code{sample=TURE}. Default to
\code{length(x)}}

\item{replace}{logical. If \code{TRUE}, sampling (in case of
\code{sample=TURE}) will be with replacement.}
}
\value{
A vector of length \code{length(.bin)} corresponding to survival
proportion at each of timing bin in \code{.bin}.
}
\description{
Cacluate survival proportion of Reaction Time (RT) data in each bin
}
\details{
This funciton is used internally to calculate the survival proportion.
}
\examples{

#generate a random RT data
set.seed(1)
sbj.x <- rnorm(n = 50, mean=300, sd=50)

#obtain a survival curve for 600 bins form 0 ms to 600 ms.
curv.x <- surv.curv(sbj.x, .bin=seq(0,600,by=1))

#plot
plot(curv.x)

}
\seealso{
\code{\link{surv.mean}}
}

